+++ install.sh
@@ -153,7 +153,12 @@ for arg in "$@"; do
 done
 
 # Welcome message
-cat $SRC_DIR/motd
+if ! ENVSUBST_LOC="$(type -p "envsubst")" || [[ -z ENVSUBST_LOC ]]; then
+  echo "envsubst not found but it's required to run this script. Try to install gettext or gettext-base package"
+  exit 1
+fi
+
+CURRENT_YEAR=$(date +"%Y") envsubst < $SRC_DIR/motd
 
 # Install PipelineWise core components
 cd $SRC_DIR
@@ -173,6 +178,7 @@ DEFAULT_CONNECTORS=(
     tap-github
     tap-slack
     tap-mixpanel
+    tap-twilio
     target-s3-csv
     target-snowflake
     target-redshift