+++ pipelinewise/cli/tap_properties.py
@@ -266,6 +266,14 @@ def get_tap_properties(tap=None, temp_dir=None):
             'default_replication_method': 'LOG_BASED',
             'default_data_flattening_max_level': 0
         },
+        'tap-twilio': {
+            'tap_config_extras': {},
+            'tap_stream_id_pattern': '{{table_name}}',
+            'tap_stream_name_pattern': '{{table_name}}',
+            'tap_catalog_argument': '--catalog',
+            'default_replication_method': 'INCREMENTAL',
+            'default_data_flattening_max_level': 0
+        },
         # Default values to use as a fallback method
         'DEFAULT': {
             'tap_config_extras': {},