+++ pipelinewise/fastsync/commons/utils.py
@@ -5,11 +5,22 @@ import os
 import logging
 from typing import Dict
 
+from typing import Dict
+
 LOGGER = logging.getLogger(__name__)
 
 SDC_EXTRACTED_AT = '_SDC_EXTRACTED_AT'
 SDC_BATCHED_AT = '_SDC_BATCHED_AT'
 SDC_DELETED_AT = '_SDC_DELETED_AT'
+
+
+class NotSelectedTableException(Exception):
+    """
+    Exception to raise when a table is not selected for resync
+    """
+    def __init__(self, table_name, selected_tables):
+        self.message = f'Cannot Resync unselected table "{table_name}"! Selected tables are: {selected_tables}'
+        super().__init__(self, self.message)
 
 
 # pylint: disable=missing-function-docstring
@@ -62,11 +73,11 @@ def tablename_to_dict(table, separator='.'):
     }
 
 
-def get_tables_from_properties(properties):
+def get_tables_from_properties(properties: Dict) -> set:
     """Get list of selected tables with schema names from properties json
     The output is used to generate list of tables to sync
     """
-    tables = []
+    tables = set()
 
     for stream in properties.get('streams', tables):
         metadata = stream.get('metadata', [])
@@ -80,10 +91,10 @@ def get_tables_from_properties(properties):
 
         if table_name and selected:
             if schema_name is not None or db_name is not None:
-                tables.append('{}.{}'.format(schema_name or db_name, table_name))
+                tables.add('{}.{}'.format(schema_name or db_name, table_name))
             else:
                 # Some tap types don't have db name nor schema name
-                tables.append(table_name)
+                tables.add(table_name)
 
     return tables
 
@@ -263,7 +274,8 @@ def save_state_file(path, table, bookmark, dbname=None):
     save_dict_to_json(path, state)
 
 
-def parse_args(required_config_keys):
+
+def parse_args(required_config_keys: Dict) -> argparse.Namespace:
     """Parse standard command-line args.
 
     --tap               Tap Config file
@@ -273,6 +285,7 @@ def parse_args(required_config_keys):
     --transform         Transformations Config file
     --tables            Tables to sync. (Separated by comma)
     --temp_dir          Directory to create temporary csv exports. Defaults to current work dir.
+    --drop_pg_slot      flag to drop or not the Postgres replication slot before starting the resync
 
     Returns the parsed args object from argparse. For each argument that
     point to JSON files (tap, state, properties, target, transform),
@@ -285,10 +298,11 @@ def parse_args(required_config_keys):
     parser.add_argument('--target', help='Target Config file', required=True)
     parser.add_argument('--transform', help='Transformations Config file')
     parser.add_argument('--tables', help='Sync only specific tables')
-    parser.add_argument('--export-dir', help='Temporary directory required for CSV exports')
     parser.add_argument('--temp_dir', help='Temporary directory required for CSV exports')
+    parser.add_argument('--drop_pg_slot', help='Drop pg replication slot before starting resync', action='store_true')
 
-    args = parser.parse_args()
+    args: argparse.Namespace = parser.parse_args()
+
     if args.tap:
         args.tap = load_json(args.tap)
 
@@ -303,14 +317,23 @@ def parse_args(required_config_keys):
     else:
         args.transform = {}
 
-    if args.tables:
-        args.tables = args.tables.split(',')
-    else:
-        args.tables = get_tables_from_properties(args.properties)
+    # get all selected tables from json schema
+    all_selected_tables = get_tables_from_properties(args.properties)
 
-    if args.temp_dir:
-        args.temp_dir = args.temp_dir
+    if args.tables:
+        # prevent duplicates
+        unique_tables_list = set(args.tables.split(','))
+
+        # check if all the given tables are actually selected
+        for table in unique_tables_list:
+            if table not in all_selected_tables:
+                raise NotSelectedTableException(table, all_selected_tables)
+
+        args.tables = unique_tables_list
     else:
+        args.tables = all_selected_tables
+
+    if not args.temp_dir:
         args.temp_dir = os.path.realpath('.')
 
     check_config(args.tap, required_config_keys['tap'])