+++ pipelinewise/fastsync/mysql_to_postgres.py
@@ -58,7 +58,7 @@ def tap_type_to_target_type(mysql_type, mysql_column_type):
         'longtext': 'CHARACTER VARYING',
         'enum': 'CHARACTER VARYING',
         'int': 'INTEGER NULL',
-        'tinyint': 'BOOLEAN' if mysql_column_type == 'tinyint(1)' else 'SMALLINT NULL',
+        'tinyint': 'BOOLEAN' if mysql_column_type and mysql_column_type.startswith('tinyint(1)') else 'SMALLINT NULL',
         'smallint': 'SMALLINT NULL',
         'mediumint': 'INTEGER NULL',
         'bigint': 'BIGINT NULL',