+++ pipelinewise/fastsync/postgres_to_postgres.py
@@ -2,11 +2,11 @@
 import os
 import sys
 import time
+import multiprocessing
+
 from functools import partial
 from argparse import Namespace
-import multiprocessing
 from typing import Union
-
 from datetime import datetime
 
 from ..logger import Logger
@@ -149,6 +149,10 @@ def main_impl():
             Pool size                      : %s
         -------------------------------------------------------
         """, args.tables, len(args.tables), pool_size)
+
+    # if internal arg drop_pg_slot is set to True, then we drop the slot before starting resync
+    if args.drop_pg_slot:
+        FastSyncTapPostgres.drop_slot(args.tap)
 
     # Create target schemas sequentially, Postgres doesn't like it running in parallel
     postgres_target = FastSyncTargetPostgres(args.target, args.transform)